Fixed all the things
Fixed: All the bugs with the last version. (% not working well, etc.)
Added: Timer to prevent bets from going in early (quite accurate, even
with leaderbored)
Added but failed: Coloring of names based on move.  We'll get to that
later (as well as some other stuff I have planned for that).